Updated kaguyami app to new format
