+++ isis/src/kaguya/apps/kaguyami2isis/kaguyami2isis.h
@@ -1,0 +1,10 @@
+#ifndef kaguyami2isis_h
+#define kaguyami2isis_h
+
+#include "UserInterface.h"
+
+namespace Isis {
+  extern void kaguyami2isis(UserInterface &ui);
+}
+
+#endif
